# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRE Loan Origination System API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardSummaryCountByStatus = { [key: string]: number };

export interface DashboardSummary {
  totalApplications: number;
  totalLoanVolume: number;
  inquiryCount: number;
  inReviewCount: number;
  closingCount: number;
  closedCount: number;
  avgDSCR: number;
  avgLTV: number;
  countByStatus?: DashboardSummaryCountByStatus;
}

export interface ActivityItem {
  id: string;
  applicationId: string;
  action: string;
  description: string;
  userName?: string;
  timestamp: string;
}

export interface PipelinePhase {
  phase: string;
  count: number;
  totalVolume: number;
}

export interface LoanApplication {
  id: string;
  status: string;
  borrowerId: string;
  propertyId: string;
  loanType: string;
  loanAmountUsd: string;
  loanTermYears?: string;
  interestType?: string;
  interestRatePct?: string;
  amortizationType?: string;
  ltvPct?: string;
  dscrRatio?: string;
  targetClosingDate?: string;
  rateType?: string;
  allInFixedRate?: string;
  proformaAdjustableAllInRate?: string;
  borrowerName?: string;
  entityName?: string;
  propertyAddress?: string;
  propertyCity?: string;
  propertyState?: string;
  propertyType?: string;
  createdAt: string;
  updatedAt: string;
}

export interface Borrower {
  id: string;
  firstName: string;
  lastName: string;
  entityName?: string;
  email?: string;
  phone?: string;
  mailingAddress?: string;
  city?: string;
  state?: string;
  zipCode?: string;
  creExperienceYears?: string;
  netWorthUsd?: string;
  liquidityUsd?: string;
  creditScore?: string;
  createdAt: string;
  updatedAt: string;
}

export interface Property {
  id: string;
  streetAddress: string;
  city: string;
  state: string;
  zipCode: string;
  propertyType: string;
  grossSqFt?: string;
  numberOfUnits?: string;
  yearBuilt?: string;
  physicalOccupancyPct?: string;
  economicOccupancyPct?: string;
  latitude?: string;
  longitude?: string;
  createdAt: string;
  updatedAt: string;
}

export interface LoanTerms {
  applicationId?: string;
  rateType?: string;
  baseRate?: string;
  fixedRateVariance?: string;
  indexName?: string;
  indexRate?: string;
  spreadOnFixed?: string;
  allInFixedRate?: string;
  adjustableRateVariance?: string;
  adjustableIndexName?: string;
  adjustableIndexRate?: string;
  spreadOnAdjustable?: string;
  proformaAdjustableAllInRate?: string;
  loanAmountUsd?: string;
  loanTermYears?: string;
  amortizationYears?: string;
  amortizationType?: string;
  ltvPct?: string;
  dscrRatio?: string;
  targetClosingDate?: string;
  interestOnlyPeriodMonths?: string;
  prepaymentPenaltyPct?: string;
  originationFeePct?: string;
  ecoaNoticeDate?: string;
  ecoaActionTaken?: string;
}

export type LoanApplicationDetail = LoanApplication & {
  borrower?: Borrower;
  property?: Property;
  loanTerms?: LoanTerms;
  commentCount?: number;
  documentCount?: number;
};

export interface CreateApplicationRequest {
  borrowerId: string;
  propertyId: string;
  loanType: string;
  loanAmountUsd: string;
  loanTermYears?: string;
  interestType?: string;
  amortizationType?: string;
  ltvPct?: string;
  dscrRatio?: string;
  targetClosingDate?: string;
}

export interface UpdateApplicationRequest {
  status?: string;
  loanType?: string;
  loanAmountUsd?: string;
  loanTermYears?: string;
  interestType?: string;
  interestRatePct?: string;
  amortizationType?: string;
  ltvPct?: string;
  dscrRatio?: string;
  targetClosingDate?: string;
}

export interface AdvanceStatusRequest {
  nextStatus: string;
}

export interface Comment {
  id: string;
  applicationId: string;
  parentCommentId?: string | null;
  body: string;
  authorName: string;
  authorSid?: string;
  createdAt: string;
  replies?: Comment[];
}

export interface CreateCommentRequest {
  body: string;
  authorName: string;
  authorSid?: string;
  parentCommentId?: string;
}

export interface Document {
  id: string;
  applicationId: string;
  fileName: string;
  fileType: string;
  fileSizeBytes?: number;
  category?: string;
  uploadedBy: string;
  uploadedAt: string;
  description?: string;
}

export interface CreateDocumentRequest {
  fileName: string;
  fileType: string;
  fileSizeBytes?: number;
  category?: string;
  uploadedBy: string;
  description?: string;
}

export interface CreateBorrowerRequest {
  firstName: string;
  lastName: string;
  entityName?: string;
  email?: string;
  phone?: string;
  mailingAddress?: string;
  city?: string;
  state?: string;
  zipCode?: string;
  creExperienceYears?: string;
  netWorthUsd?: string;
  liquidityUsd?: string;
  creditScore?: string;
}

export interface CreatePropertyRequest {
  streetAddress: string;
  city: string;
  state: string;
  zipCode: string;
  propertyType: string;
  grossSqFt?: string;
  numberOfUnits?: string;
  yearBuilt?: string;
  physicalOccupancyPct?: string;
  economicOccupancyPct?: string;
  latitude?: string;
  longitude?: string;
}

export interface AppUser {
  id: string;
  sid: string;
  firstName: string;
  lastName: string;
  email?: string;
  role: string;
  department?: string;
  active?: boolean;
  createdAt: string;
}

export interface CreateUserRequest {
  sid: string;
  firstName: string;
  lastName: string;
  email?: string;
  role: string;
  department?: string;
}

export interface AmortizationRequest {
  loanAmount: number;
  annualRatePct: number;
  termMonths: number;
  amortizationMonths: number;
  interestOnlyMonths?: number;
}

export interface AmortizationRow {
  period: number;
  payment: number;
  principal: number;
  interest: number;
  balance: number;
}

export interface AmortizationResult {
  monthlyPayment: number;
  totalInterest: number;
  totalPaid: number;
  balloonPayment?: number;
  schedule: AmortizationRow[];
}

export interface ApiError {
  message: string;
  code?: string;
}

export type GetRecentActivityParams = {
  limit?: number;
};

export type ListApplicationsParams = {
  status?: string;
  phase?: ListApplicationsPhase;
  search?: string;
  rateType?: string;
};

export type ListApplicationsPhase =
  (typeof ListApplicationsPhase)[keyof typeof ListApplicationsPhase];

export const ListApplicationsPhase = {
  Inquiry: "Inquiry",
  ICR: "ICR",
  Application: "Application",
  FCR: "FCR",
  Closing: "Closing",
  Disposed: "Disposed",
} as const;

export type ListBorrowersParams = {
  search?: string;
};

export type ListPropertiesParams = {
  search?: string;
  propertyType?: string;
};