# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: CRE Loan Origination System API
servers:
  - url: /java-api
    description: Spring Boot API
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: propertyType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Property list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

  # ── Users ──────────────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"

  # ── Amortization Calculator ────────────────────────────────────────────────
...
components:
  schemas:
...
    Property:
      type: object
      required: [id, streetAddress, city, state, zipCode, propertyType, createdAt, updatedAt]
      properties:
        id:
          type: string
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        propertyType:
          type: string
        grossSqFt:
          type: string
        numberOfUnits:
          type: string
        yearBuilt:
          type: string
        physicalOccupancyPct:
          type: string
        economicOccupancyPct:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePropertyRequest:
      type: object
      required: [streetAddress, city, state, zipCode, propertyType]
      properties:
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        propertyType:
          type: string
        grossSqFt:
          type: string
        numberOfUnits:
          type: string
        yearBuilt:
          type: string
        physicalOccupancyPct:
          type: string
        economicOccupancyPct:
          type: string
        latitude:
          type: string
        longitude:
          type: string

    # ── AppUser ───────────────────────────────────────────────────────────────
    AppUser:
      type: object
      required: [id, sid, firstName, lastName, role, createdAt]
      properties:
        id:
          type: string
        sid:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [sid, firstName, lastName, role]
      properties:
        sid:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string

    # ── Amortization ──────────────────────────────────────────────────────────
...