# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRE Loan Origination System API
 * OpenAPI spec version: 1.0.0
 */

export interface Property {
  id: string;
  streetAddress: string;
  city: string;
  state: string;
  zipCode: string;
  propertyType: string;
  grossSqFt?: string;
  numberOfUnits?: string;
  yearBuilt?: string;
  physicalOccupancyPct?: string;
  economicOccupancyPct?: string;
  latitude?: string;
  longitude?: string;
  createdAt: Date;
  updatedAt: Date;
}